Made changes to tours.jsx
